+++ LICENSE.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 58099d1e589583e45942e2531792af72
-folderAsset: yes
+guid: ebf9c2a61846e8e4989be1d4a6354ccd
 DefaultImporter:
   externalObjects: {}
   userData:
