+++ README.md.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 713ed17dd84ce2042b0da051953f62a0
-folderAsset: yes
-DefaultImporter:
+guid: 85f019d4eb452094caea94c65ef6f29f
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
